Revert "zenject wip"
This reverts commit 3fce1d35e282af862e6b1beb60855cfb368368c8.

+++ Assets/_SunsetSystems/Entities/Characters/Data/CreatureData.cs
@@ -24,7 +24,7 @@ namespace SunsetSystems.Entities.Characters
         [ES3Serializable]
         public EquipmentData Equipment;
         public bool UseEquipmentPreset;
-        public int Money;
+        public float Money;
 
         public CreatureData(CreatureConfig config)
         {
