+++ Assets/_SunsetSystems/Entities/Characters/Scripts/InventoryConfig.cs
@@ -13,6 +13,6 @@ namespace SunsetSystems.Entities.Characters
         [field: SerializeField]
         public List<InventoryEntry> Inventory { get; private set; }
         [field: SerializeField]
-        public int Money { get; private set; }
+        public float Money { get; private set; }
     }
 }
